backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
langchain==0.1.0
langchain-community==0.0.16
langchain-core==0.1.18
langchain-openai==0.0.5
langchain-pinecone==0.0.1
pinecone-client==2.2.4
openai==1.12.0
python-jose==3.3.0
mysql-connector-python==8.0.26
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2